+++ setup.py
@@ -6,15 +6,15 @@ with open("README.md", "r") as fh:
     long_description = fh.read()
 
 setuptools.setup(
-    name='spg-topdown-traits',
-    version='0.4.7',
+    name='spg',
+    version='0.4.8',
     description='Extract geometric traits from top-view images of plants.',
     long_description=long_description,
     long_description_content_type="text/markdown",
     author='Suxing Liu',
     author_email='suxing.liu@uga.edu',
     license='BSD-3-Clause',
-    url='https://github.com/Computational-Plant-Science/spg-topdown-traits',
+    url='https://github.com/Computational-Plant-Science/spg',
     packages=setuptools.find_packages(),
     include_package_data=True,
     entry_points={
