+++ spg/luminosity.py
